refactor: Split Python client.py into focused sub-modules
- Created transport.py: HTTP transport layer with session management, retry logic, and response handling
- Created endpoints.py: API endpoint registry for centralized module initialization
- Converted client.py to client/__init__.py: High-level facade that delegates to transport and endpoints
- Maintained backward compatibility with existing API

Related to issue #5
